Refactor API graph ops
